Contraseña encriptada

+++ Programacion/bancoDeImagenesBK/routes/empleados.js
@@ -2,8 +2,10 @@ var express = require('express');
 var router = express.Router();
 const pool = require('../database');
 const passport = require('passport');
+const { check, validationResult } = require('express-validator');
+const bcrypt = require('bcrypt');
 
-/* GET users listing. */
+/* METODOS PARA CRUD*/
 router.get('/', async(req, res, next) => {
     const empleado = await pool.query('SELECT * FROM empleados');
     res.json(empleado);
@@ -19,8 +21,27 @@ router.get('/:usuario', async(req, res, next) => {
     res.status(404).json({ text: 'Registro no existente' });
 });
 
-router.post('/', async(req, res) => {
-    const { nombre, apellidos, fechaN, puesto, fechaIngreso, usuario, contraseña } = req.body;
+router.post('/', [
+    check('nombre').isLength({ min: 1 }),
+    check('apellidos').isLength({ min: 3, max: 60 }),
+    check('fechaN').isLength({ min: 3, max: 10 }),
+    check('puesto').isLength({ min: 3 }),
+    check('fechaIngreso').isLength({ min: 3, max: 10 }),
+    check('usuario').isLength({ min: 3 }),
+    check('password').isLength({ min: 5 })
+], async(req, res) => {
+
+
+    const errores = validationResult(req);
+    if (!errores.isEmpty()) {
+        return res.status(422).json({ error: errores.array() });
+    }
+
+    const { nombre, apellidos, fechaN, puesto, fechaIngreso, usuario } = req.body;
+
+    const salt = await bcrypt.genSalt(10);
+    const password = await bcrypt.hash(req.body.password, salt);
+
     const newempleado = {
         nombre,
         apellidos,
@@ -28,7 +49,7 @@ router.post('/', async(req, res) => {
         puesto,
         fechaIngreso,
         usuario,
-        contraseña
+        password
     };
     await pool.query('INSERT INTO empleados set ?', [newempleado]);
 
@@ -38,36 +59,81 @@ router.post('/', async(req, res) => {
 });
 
 router.post('iniciosesion', (req, res) => {
-    passport.authenticate('loca.signup', {
+    passport.authenticate('local.signup', {
         successRedirect: '/inicio',
-        failureRedirect: 'signup'
+        failureRedirect: '/iniciosesion',
+        failureFlash: true
     })
     console.log('iniciar sesion');
 });
 
-router.put('/:usuario', async(req, res) => {
+router.put('/:usuario', [
+    check('nombre').isLength({ min: 1 }),
+    check('apellidos').isLength({ min: 3, max: 60 }),
+    check('fechaN').isLength({ min: 3, max: 10 }),
+    check('puesto').isLength({ min: 3 }),
+    check('fechaIngreso').isLength({ min: 3, max: 10 }),
+    check('usuario').isLength({ min: 3 }),
+    check('password').isLength({ min: 5 })
+], async(req, res) => {
+
+    const errores = validationResult(req);
+    if (!errores.isEmpty()) {
+        return res.status(422).json({ error: errores.array() });
+    }
+
     const { usuario } = req.params;
-    const { nombre, apellidos, fechaN, puesto, fechaIngreso, contraseña } = req.body;
-    const empleado = {
-        nombre,
-        apellidos,
-        fechaN,
-        puesto,
-        fechaIngreso,
-        usuario,
-        contraseña
-    };
-    await pool.query('UPDATE empleados SET ? WHERE usuario = ?', [empleado, usuario]);
-    res.json({ message: 'El registro fue actualizado' });
-    console.log(req.body);
+    const empleado = await pool.query('SELECT * FROM empleados WHERE usuario = ?', [usuario]);
+    if (empleado.length > 0) {
+
+
+        const { nombre, apellidos, fechaN, puesto, fechaIngreso } = req.body;
+
+        const salt = await bcrypt.genSalt(10);
+        const password = await bcrypt.hash(req.body.password, salt);
+
+        const newempleado = {
+            nombre,
+            apellidos,
+            fechaN,
+            puesto,
+            fechaIngreso,
+            usuario,
+            password
+        };
+
+        await pool.query('UPDATE empleados SET ? WHERE usuario = ?', [newempleado, usuario]);
+
+        res.json({ message: 'El registro fue actualizado' });
+
+        console.log(req.body);
+    } else {
+        res.status(400).json({ text: 'El usuario no fue encontrado' });
+    }
+
+
 });
 
-router.delete('/:usuario', async(req, res) => {
+router.delete('/:usuario', [
+    check('usuario').isLength({ min: 3 })
+], async(req, res) => {
+
+    errores = validationResult(req);
+    if (!errores.isEmpty()) {
+        return res.status(422).json({ error: errores.array() });
+    }
+
     const { usuario } = req.params;
-    await pool.query('DELETE FROM empleados WHERE usuario = ?', [usuario]);
-    res.json({ message: 'El registro fue eliminado' });
-    console.log(req.body);
+    const empleado = await pool.query('SELECT * FROM empleados WHERE usuario = ?', [usuario]);
+    if (empleado.length > 0) {
+        await pool.query('DELETE FROM empleados WHERE usuario = ?', [usuario]);
+        res.json({ message: 'El registro fue eliminado' });
+        console.log(req.body);
+    } else {
+        res.status(400).json({ text: 'El usuario no existe' });
+    }
 });
+
